target computation time and new compile definition

+++ playground/c++/src/ogdf_maxsat.cpp
@@ -226,7 +226,8 @@ void solveEvalMaxSat(Instance& I) {
 	std::ofstream f(filename);
 #endif
 
-	solver.setTargetComputationTime(std::min(10 * 60, I.G.numberOfNodes() / 10 + 1));
+	solver.setTargetComputationTime(
+			std::min((double)(10 * 60), (double)I.G.numberOfNodes() / (double)10 + 1));
 	std::cout.setstate(std::ios::failbit); // https://stackoverflow.com/a/8246430
 	bool solved = solver.solve();
 	std::cout.clear();
